requirements.txt
asgiref==3.7.2
Django==5.0.2
sqlparse==0.4.4
psycopg2-binary==2.9.9
djangorestframework==3.15.1
